+++ packages/xchain-cosmos/src/client.ts
@@ -228,6 +228,15 @@ class Client implements CosmosClient, XChainClient {
       average: baseAmount(0, 6),
     })
   }
+
+  getDefaultFees = (): Fees => {
+    return {
+      type: 'base',
+      fast: baseAmount(750, 6),
+      fastest: baseAmount(2500, 6),
+      average: baseAmount(0, 6),
+    }
+  }
 }
 
 export { Client }