+++ packages/xchain-cosmos/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v.x.x.x (2020-XX-XX)
+
+### Update
+
+- Update dependencies
+- Move `cosmos-client` to `dependencies`
+
 # v.0.4.2 (2020-11-23)
 
 ### Fix imports